Create Windows Security Log Event ID 4688
Event 4688 documents each program that is executed, who the program ran as and the process that started this process. When you start a program you are creating a "process" that stays open until the program exits.  This process is identified by the Process ID:.  You can correlate this event to other events by Process ID to determine what the program did while it ran and when it exited (event 4689). This query finds top processes executed by the command line.